+++ test/Fixture/DefinitionProvider/ImplementsDefinitionProvider/UserDefinitionProvider.php
@@ -11,13 +11,13 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definitions\CanNotBeAutoloaded;
+namespace Ergebnis\FactoryBot\Test\Fixture\DefinitionProvider\ImplementsDefinitionProvider;
 
-use Ergebnis\FactoryBot\Definition;
+use Ergebnis\FactoryBot\EntityDefinitionProvider;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 
-final class UserDefinition implements Definition
+final class UserDefinitionProvider implements EntityDefinitionProvider
 {
     public function accept(FixtureFactory $fixtureFactory): void
     {