+++ test/Fixture/DefinitionProvider/ImplementsDefinitionProviderButCanNotBeInstantiated/RepositoryDefinitionProvider.php
@@ -11,16 +11,20 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definitions\ImplementsDefinition;
+namespace Ergebnis\FactoryBot\Test\Fixture\DefinitionProvider\ImplementsDefinitionProviderButCanNotBeInstantiated;
 
-use Ergebnis\FactoryBot\Definition;
+use Ergebnis\FactoryBot\EntityDefinitionProvider;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 
-final class UserDefinition implements Definition
+final class RepositoryDefinitionProvider implements EntityDefinitionProvider
 {
+    private function __construct()
+    {
+    }
+
     public function accept(FixtureFactory $fixtureFactory): void
     {
-        $fixtureFactory->define(Fixture\FixtureFactory\Entity\User::class);
+        $fixtureFactory->define(Fixture\FixtureFactory\Entity\Repository::class);
     }
 }