+++ test/Fixture/DefinitionProvider/ImplementsDefinitionProviderButIsAbstract/RepositoryDefinitionProvider.php
@@ -11,14 +11,15 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definitions\DoesNotImplementDefinition;
+namespace Ergebnis\FactoryBot\Test\Fixture\DefinitionProvider\ImplementsDefinitionProviderButIsAbstract;
 
+use Ergebnis\FactoryBot\EntityDefinitionProvider;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 
-final class RepositoryDefinition
+abstract class RepositoryDefinitionProvider implements EntityDefinitionProvider
 {
-    public function accept(FixtureFactory $fixtureFactory): void
+    final public function accept(FixtureFactory $fixtureFactory): void
     {
         $fixtureFactory->define(Fixture\FixtureFactory\Entity\Repository::class);
     }