+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For a full diff see [`0.1.0...main`][0.1.0...main].
 * Renamed `Number` to `Count` ([#309]), by [@localheinz]
 * Split `FixtureFactory::persistOnGet()` into `FixtureFactory::persistAfterCreate()` and `FixtureFactory::doNotPersistAfterCreate()` ([#311]), by [@localheinz]
 * Merged `Definitions` into `FixtureFactory::load()` ([#312]), by [@localheinz]
+* Renamed `Definition` to `EntityDefinitionProvider` ([#314]), by [@localheinz]
 
 ### Fixed
 
@@ -152,5 +153,6 @@ For a full diff see [`fa9c564...0.1.0`][fa9c564...0.1.0].
 [#309]: https://github.com/ergebnis/factory-bot/pull/309
 [#311]: https://github.com/ergebnis/factory-bot/pull/311
 [#312]: https://github.com/ergebnis/factory-bot/pull/312
+[#314]: https://github.com/ergebnis/factory-bot/pull/314
 
 [@localheinz]: https://github.com/localheinz